Safari-611.1.10/LayoutTests/media/modern-media-controls/playback-support/playback-support-autoplay-expected.txt

Testing the PlaybackSupport behavior when the video autoplay.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media is playing
PASS mediaController.controls.playPauseButton.playing is true

PASS successfullyParsed is true

TEST COMPLETE
